% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_obama.R
\name{is_obama}
\alias{is_obama}
\title{Is this an Obama election year?}
\usage{
is_obama(num)
}
\arguments{
\item{num}{a number, or a string that contains a number}
}
\value{
TRUE or FALSE on whether or not it is an obama election year
}
\description{
Is this an Obama election year?
}
\examples{
is_obama(2008) # returns TRUE
is_obama(2012) # returns TRUE
is_obama(2016) # returns FALSE
is_obama("His first election was in 2008") # returns TRUE

\dontrun{
is_obama(mtcars)
}

}
